Adding macros to push pinning roots (#43)
Supporting pushing roots to the shadow stack that are not transitively pinned (only the root object is pinned), and therefore should enable moving more objects.

For backwards compatibility the original functions will be used to represent transitively pinned roots, whereas the `*_NO_TPIN` version will represent roots in which only the root object needs to be pinned. Therefore, to support moving move objects the goal is to replace as much as possible of the original functions with their `_NO_TPIN` counterpart.